+++ test/KBaseDataObjectToFileUtils_server_test.py
@@ -77,6 +77,128 @@ class KBaseDataObjectToFileUtilsTest(unittest.TestCase):
     def getContext(self):
         return self.__class__.ctx
 
+    # call this method to get the WS object info of a Genome
+    #   (will upload the example data if this is the first time the method is called during tests)
+    def getGenomeInfo(self, genome_basename, item_i=0):
+        if hasattr(self.__class__, 'genomeInfo_list'):
+            try:
+                info = self.__class__.genomeInfo_list[item_i]
+                name = self.__class__.genomeName_list[item_i]
+                if info != None:
+                    if name != genome_basename:
+                        self.__class__.genomeInfo_list[item_i] = None
+                        self.__class__.genomeName_list[item_i] = None
+                    else:
+                        return info
+            except:
+                pass
+
+        # 1) transform genbank to kbase genome object and upload to ws
+        shared_dir = "/kb/module/work/tmp"
+        genome_data_file = 'data/genomes/'+genome_basename+'.gbff.gz'
+        genome_file = os.path.join(shared_dir, os.path.basename(genome_data_file))
+        shutil.copy(genome_data_file, genome_file)
+
+        SERVICE_VER = 'release'
+        #SERVICE_VER = 'dev'
+        GFU = GenomeFileUtil(os.environ['SDK_CALLBACK_URL'],
+                             token=self.getContext()['token'],
+                             service_ver=SERVICE_VER
+                         )
+        print ("UPLOADING genome: "+genome_basename+" to WORKSPACE "+self.getWsName()+" ...")
+        genome_upload_result = GFU.genbank_to_genome({'file': {'path': genome_file },
+                                                      'workspace_name': self.getWsName(),
+                                                      'genome_name': genome_basename
+                                                  })
+#                                                  })[0]
+        pprint(genome_upload_result)
+        genome_ref = genome_upload_result['genome_ref']
+        new_obj_info = self.getWsClient().get_object_info_new({'objects': [{'ref': genome_ref}]})[0]
+
+        # 2) store it
+        if not hasattr(self.__class__, 'genomeInfo_list'):
+            self.__class__.genomeInfo_list = []
+            self.__class__.genomeName_list = []
+        for i in range(item_i+1):
+            try:
+                assigned = self.__class__.genomeInfo_list[i]
+            except:
+                self.__class__.genomeInfo_list.append(None)
+                self.__class__.genomeName_list.append(None)
+
+        self.__class__.genomeInfo_list[item_i] = new_obj_info
+        self.__class__.genomeName_list[item_i] = genome_basename
+        return new_obj_info
+
+    # call this method to get the WS object info of an AnnotatedMetagenomeAssembly
+    #   (will upload the example data if this is the first time the method is called during tests)
+    def getAMAInfo(self, ama_basename, item_i=0):
+        if hasattr(self.__class__, 'amaInfo_list'):
+            try:
+                info = self.__class__.amaInfo_list[item_i]
+                name = self.__class__.amaName_list[item_i]
+                if info != None:
+                    if name != ama_basename:
+                        self.__class__.amaInfo_list[item_i] = None
+                        self.__class__.amaName_list[item_i] = None
+                    else:
+                        return info
+            except:
+                pass
+
+        # 1) transform GFF+FNA to kbase AMA object and upload to ws
+        shared_dir = "/kb/module/work/tmp"
+        ama_gff_srcfile = 'data/amas/'+ama_basename+'.gff'
+        ama_fna_srcfile = 'data/amas/'+ama_basename+'.fa'
+        ama_gff_dstfile = os.path.join(shared_dir, os.path.basename(ama_gff_srcfile))
+        ama_fna_dstfile = os.path.join(shared_dir, os.path.basename(ama_fna_srcfile))
+        shutil.copy(ama_gff_srcfile, ama_gff_dstfile)
+        shutil.copy(ama_fna_srcfile, ama_fna_dstfile)
+
+        try:
+            SERVICE_VER = 'release'
+            #SERVICE_VER = 'dev'
+            GFU = GenomeFileUtil(os.environ['SDK_CALLBACK_URL'],
+                                 token=self.getContext()['token'],
+                                 service_ver=SERVICE_VER
+            )
+        except:
+            raise ValueError ("unable to obtain GenomeFileUtil client")
+        print ("UPLOADING AMA: "+ama_basename+" to WORKSPACE "+self.getWsName()+" ...")
+        ama_upload_params = {
+            "workspace_name": self.getWsName(),
+            "genome_name": ama_basename,
+            "fasta_file": {"path": ama_fna_dstfile},
+            "gff_file": {"path": ama_gff_dstfile},
+            "source": "GFF",
+            "scientific_name": "TEST AMA",
+            "generate_missing_genes": "True"
+        }        
+        try:
+            ama_upload_result = GFU.fasta_gff_to_metagenome(ama_upload_params)
+        except:
+            raise ValueError("unable to upload test AMA data object")
+        print ("AMA UPLOADED")
+        pprint(ama_upload_result)
+
+        ama_ref = ama_upload_result['metagenome_ref']
+        new_obj_info = self.getWsClient().get_object_info_new({'objects': [{'ref': ama_ref}]})[0]
+
+        # 2) store it
+        if not hasattr(self.__class__, 'amaInfo_list'):
+            self.__class__.amaInfo_list = []
+            self.__class__.amaName_list = []
+        for i in range(item_i+1):
+            try:
+                assigned = self.__class__.amaInfo_list[i]
+            except:
+                self.__class__.amaInfo_list.append(None)
+                self.__class__.amaName_list.append(None)
+
+        self.__class__.amaInfo_list[item_i] = new_obj_info
+        self.__class__.amaName_list[item_i] = ama_basename
+        return new_obj_info
+    
     # NOTE: According to Python unittest naming rules test method names should start from 'test'.
 
 
@@ -134,16 +256,22 @@ class KBaseDataObjectToFileUtilsTest(unittest.TestCase):
     ##
     # HIDE @unittest.skip ('skipping test_KBaseDataObjectToFileUtils_GenomeToFASTA_01()')
     def test_KBaseDataObjectToFileUtils_GenomeToFASTA_01(self):
-
+        [OBJID_I, NAME_I, TYPE_I, SAVE_DATE_I, VERSION_I, SAVED_BY_I, WSID_I, WORKSPACE_I, CHSUM_I, SIZE_I, META_I] = range(11)  # object_info tuple
+        """
         reference_prok_genomes_WS = 'ReferenceDataManager'  # PROD and CI
         genome_ref_1 = reference_prok_genomes_WS+'/GCF_001566335.1/1'  # E. coli K-12 MG1655
+        """
+        genomeInfo_0 = self.getGenomeInfo('GCF_001566335.1_ASM156633v1_genomic', 0)  # E. coli K-12 MG1655
+        genome_ref_0 = '/'.join([str(genomeInfo_0[WSID_I]),
+                                 str(genomeInfo_0[OBJID_I]),
+                                 str(genomeInfo_0[VERSION_I])])        
 
         output_dir = os.path.join(self.scratch,'fasta_out.'+str(uuid.uuid4()))
         if not os.path.exists(output_dir):
             os.makedirs(output_dir)
 
         parameters = {
-                'genome_ref':          genome_ref_1,
+                'genome_ref':          genome_ref_0,
                 'file':                'test_genome.fasta',
                 'dir':                 output_dir,
                 'console':             [],
@@ -166,17 +294,34 @@ class KBaseDataObjectToFileUtilsTest(unittest.TestCase):
     ##
     # HIDE @unittest.skip ('skipping test_KBaseDataObjectToFileUtils_GenomeSetToFASTA_01()')
     def test_KBaseDataObjectToFileUtils_GenomeSetToFASTA_01(self):
+        [OBJID_I, NAME_I, TYPE_I, SAVE_DATE_I, VERSION_I, SAVED_BY_I, WSID_I, WORKSPACE_I, CHSUM_I, SIZE_I, META_I] = range(11)  # object_info tuple
 
+        """
         reference_prok_genomes_WS = 'ReferenceDataManager'  # PROD and CI
         genome_ref_1 = reference_prok_genomes_WS+'/GCF_001566335.1/1'  # E. coli K-12 MG1655
         genome_ref_2 = reference_prok_genomes_WS+'/GCF_000021385.1/1'  # D. vulgaris str. 'Miyazaki F'
         genome_ref_3 = reference_prok_genomes_WS+'/GCF_900129775.1/1'  # Halobaculum gomorrense (16 contigs)
+        """
+        
+        genomeInfo_0 = self.getGenomeInfo('GCF_001566335.1_ASM156633v1_genomic', 0)  # E. coli K-12 MG1655
+        genomeInfo_1 = self.getGenomeInfo('GCF_000021385.1_ASM2138v1_genomic', 1)    # D. vulgaris str. 'Miyazaki F
+        genomeInfo_2 = self.getGenomeInfo('GCF_001721825.1_ASM172182v1_genomic', 2)  # Pseudomonas aeruginosa
+        genome_ref_0 = '/'.join([str(genomeInfo_0[WSID_I]),
+                                 str(genomeInfo_0[OBJID_I]),
+                                 str(genomeInfo_0[VERSION_I])])
+        genome_ref_1 = '/'.join([str(genomeInfo_1[WSID_I]),
+                                 str(genomeInfo_1[OBJID_I]),
+                                 str(genomeInfo_1[VERSION_I])])
+        genome_ref_2 = '/'.join([str(genomeInfo_2[WSID_I]),
+                                 str(genomeInfo_2[OBJID_I]),
+                                 str(genomeInfo_2[VERSION_I])])        
+        
         genome_id_feature_id_delim = '.f:'
 
         genomeSet_obj = { 'description': 'test genomeSet',
-                          'elements': { 'genome_1': { 'ref': genome_ref_1 },
-                                        'genome_2': { 'ref': genome_ref_2 },
-                                        'genome_3': { 'ref': genome_ref_3 }
+                          'elements': { 'genome_0': { 'ref': genome_ref_0 },
+                                        'genome_1': { 'ref': genome_ref_1 },
+                                        'genome_2': { 'ref': genome_ref_2 }
                                     }
                         }
         provenance = [{}]
@@ -192,7 +337,6 @@ class KBaseDataObjectToFileUtilsTest(unittest.TestCase):
                 }
             ]})[0]
 
-        [OBJID_I, NAME_I, TYPE_I, SAVE_DATE_I, VERSION_I, SAVED_BY_I, WSID_I, WORKSPACE_I, CHSUM_I, SIZE_I, META_I] = range(11)  # object_info tuple
         genomeSet_ref = str(genomeSet_info[WSID_I])+'/'+str(genomeSet_info[OBJID_I])+'/'+str(genomeSet_info[VERSION_I])
 
         output_dir = os.path.join(self.scratch,'fasta_out.'+str(uuid.uuid4()))
@@ -224,26 +368,51 @@ class KBaseDataObjectToFileUtilsTest(unittest.TestCase):
     ##
     # HIDE @unittest.skip ('skipping test_KBaseDataObjectToFileUtils_FeatureSetToFASTA_01()')
     def test_KBaseDataObjectToFileUtils_FeatureSetToFASTA_01(self):
+        [OBJID_I, NAME_I, TYPE_I, SAVE_DATE_I, VERSION_I, SAVED_BY_I, WSID_I, WORKSPACE_I, CHSUM_I, SIZE_I, META_I] = range(11)  # object_info tuple
 
+        """
         reference_prok_genomes_WS = 'ReferenceDataManager'  # PROD and CI
         genome_ref_1 = reference_prok_genomes_WS+'/GCF_001566335.1/1'  # E. coli K-12 MG1655
         genome_ref_2 = reference_prok_genomes_WS+'/GCF_000021385.1/1'  # D. vulgaris str. 'Miyazaki F'
         genome_ref_3 = reference_prok_genomes_WS+'/GCF_900129775.1/1'  # Halobaculum gomorrense (16 contigs)
+        """
+        genomeInfo_0 = self.getGenomeInfo('GCF_001566335.1_ASM156633v1_genomic', 0)  # E. coli K-12 MG1655
+        genomeInfo_1 = self.getGenomeInfo('GCF_000021385.1_ASM2138v1_genomic', 1)    # D. vulgaris str. 'Miyazaki F
+        genomeInfo_2 = self.getGenomeInfo('GCF_001721825.1_ASM172182v1_genomic', 2)  # Pseudomonas aeruginosa
+        genome_ref_0 = '/'.join([str(genomeInfo_0[WSID_I]),
+                                 str(genomeInfo_0[OBJID_I]),
+                                 str(genomeInfo_0[VERSION_I])])
+        genome_ref_1 = '/'.join([str(genomeInfo_1[WSID_I]),
+                                 str(genomeInfo_1[OBJID_I]),
+                                 str(genomeInfo_1[VERSION_I])])
+        genome_ref_2 = '/'.join([str(genomeInfo_2[WSID_I]),
+                                 str(genomeInfo_2[OBJID_I]),
+                                 str(genomeInfo_2[VERSION_I])])        
+        
+
+        amaInfo_0 = self.getAMAInfo("test_ama", 0)
+        ama_ref_0 = '/'.join([str(amaInfo_0[WSID_I]),
+                              str(amaInfo_0[OBJID_I]),
+                              str(amaInfo_0[VERSION_I])])
+
         genome_id_feature_id_delim = '.f:'
         feature_id_1 = 'AWN69_RS07145'
         feature_id_2 = 'DVMF_RS00005'
-        feature_id_3 = 'BUE16_RS15805'
-
+        feature_id_3 = 'A6701_RS00005'
+        ama_feature_id_1 = '4_27'
+        
         featureSet_obj = { 'description': 'test genomeSet',
                            'element_ordering': [
                                feature_id_1,
                                feature_id_2,
-                               feature_id_3
+                               feature_id_3,
+                               ama_feature_id_1
                            ],
                            'elements': { 
-                               feature_id_1: [genome_ref_1],
-                               feature_id_2: [genome_ref_2],
-                               feature_id_3: [genome_ref_3]
+                               feature_id_1: [genome_ref_0],
+                               feature_id_2: [genome_ref_1],
+                               feature_id_3: [genome_ref_2],
+                               ama_feature_id_1: [ama_ref_0]
                            }
                         }
         provenance = [{}]
@@ -259,7 +428,6 @@ class KBaseDataObjectToFileUtilsTest(unittest.TestCase):
                 }
             ]})[0]
 
-        [OBJID_I, NAME_I, TYPE_I, SAVE_DATE_I, VERSION_I, SAVED_BY_I, WSID_I, WORKSPACE_I, CHSUM_I, SIZE_I, META_I] = range(11)  # object_info tuple
         featureSet_ref = str(featureSet_info[WSID_I])+'/'+str(featureSet_info[OBJID_I])+'/'+str(featureSet_info[VERSION_I])
 
         output_dir = os.path.join(self.scratch,'fasta_out.'+str(uuid.uuid4()))
@@ -293,48 +461,23 @@ class KBaseDataObjectToFileUtilsTest(unittest.TestCase):
     ##
     # HIDE @unittest.skip ('skipping test_KBaseDataObjectToFileUtils_AnnotatedMetagenomeAssemblyToFASTA_01()')
     def test_KBaseDataObjectToFileUtils_AnnotatedMetagenomeAssemblyToFASTA_01(self):
+        [OBJID_I, NAME_I, TYPE_I, SAVE_DATE_I, VERSION_I, SAVED_BY_I, WSID_I, WORKSPACE_I, CHSUM_I, SIZE_I, META_I] = range(11)  # object_info tuple
+
         # upload test data
-        ama_name = "ama_test.AMA"
         ama_feature_cnt = 888
-        ama_contigs_file_src = "data/AnnotatedMetagenomeAssembly/ama_contigs.fasta"
-        ama_genes_file_src   = "data/AnnotatedMetagenomeAssembly/ama_genes.gff"
-        shared_dir = "/kb/module/work/tmp"
-        ama_contigs_file_upload = os.path.join (shared_dir, os.path.basename(ama_contigs_file_src))
-        ama_genes_file_upload = os.path.join (shared_dir, os.path.basename(ama_genes_file_src))
-        shutil.copy (ama_contigs_file_src, ama_contigs_file_upload)
-        shutil.copy (ama_genes_file_src, ama_genes_file_upload)
 
-        ama_upload_params = {
-            "workspace_name": self.getWsName(),
-            "genome_name": ama_name,
-            "fasta_file": {"path": ama_contigs_file_upload},
-            "gff_file": {"path": ama_genes_file_upload},
-            "source": "GFF",
-            "scientific_name": "TEST AMA",
-            "generate_missing_genes": "True"
-        }
-        try:
-            SERVICE_VER = 'dev'
-            GFU = GenomeFileUtil(os.environ['SDK_CALLBACK_URL'],
-                                 token=self.getContext()['token'],
-                                 service_ver=SERVICE_VER
-                             )            
-
-            print ("UPLOADING AMA: "+ama_name+" to WORKSPACE "+self.getWsName()+" ...")
-            ama_upload_result = GFU.fasta_gff_to_metagenome (ama_upload_params)
-        except:
-            raise ValueError("unable to upload test AMA data object")        
-        pprint (ama_upload_result)
-        ama_ref = ama_upload_result['metagenome_ref']
-            
-
+        amaInfo_0 = self.getAMAInfo("test_ama", 0)
+        ama_ref_0 = '/'.join([str(amaInfo_0[WSID_I]),
+                              str(amaInfo_0[OBJID_I]),
+                              str(amaInfo_0[VERSION_I])])
+        
         # get protein fastas
         output_dir = os.path.join(self.scratch,'fasta_out.'+str(uuid.uuid4()))
         if not os.path.exists(output_dir):
             os.makedirs(output_dir)
 
         parameters = {
-                'ama_ref':             ama_ref,
+                'ama_ref':             ama_ref_0,
                 'file':                'test_ama_proteins.fasta',
                 'dir':                 output_dir,
                 'console':             [],