+++ RELEASE_NOTES.md
@@ -1,7 +1,9 @@
 ### Version 1.0.0
 __Changes__
 - handle gene annotation functions from Genomes >= 9.0 (Features >= 3.0)
+- support FeatureSets with AnnotatedMetagenomeAssembly features
 - added Github Actions testing
+- made all test data uploaded
 
 ### Version 0.0.9
 __Changes__